// backend/src/lib/api-zod/src/generated/api.ts
export const CreateSplitBody = zod.object({
  contestant1Id: zod.number(),
  contestant2Id: zod.number(),
});

/**
 * @summary Break a split
 */
export const DeleteSplitParams = zod.object({
  id: zod.coerce.number(),
});